renamed test files added to gitignore and requirements files
